requirements.txt
streamlit
streamlit-webrtc
tensorflow
opencv-python
pillow
av
pillow
mtcnn